Resource_Server/resource_server/property/serializers.py
from rest_framework import serializers
from db_models.models import Property


class PropertySerializer(serializers.ModelSerializer):
    class Meta:
        model = Property
        fields = '__all__'