Resource_Server/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.9.1
oauthlib==3.1.1
psycopg2==2.7.7
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.5
wrapt==1.12.1
